requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
langgraph
wikipedia
python-dotenv